requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-channels==0.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
idna==2.10
oauthlib==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
